chore: split icon-wrapper
